LayoutTests/inspector/debugger/break-on-exception-expected.txt

CONSOLE MESSAGE: throwing TestError
CONSOLE MESSAGE: TestError
CONSOLE MESSAGE: testFinally
CONSOLE MESSAGE: throwing TestError
CONSOLE MESSAGE: finally
CONSOLE MESSAGE: TestError
CONSOLE MESSAGE: testThrowingThruNativeCode
CONSOLE MESSAGE: throwing TestError
CONSOLE MESSAGE: TestError
CONSOLE MESSAGE: testCatch
CONSOLE MESSAGE: throwing TestError
CONSOLE MESSAGE: catch TestError
CONSOLE MESSAGE: DONE
Checking pause locations when pausing on all exceptions.


== Running test suite: BreakOnAnyException
-- Running test case: BreakOnAnyException.UncaughtExceptionBasic
PASS: Should pause for exception.
PAUSE AT doThrow:8:22
Uncaught exception in test page: TestError [break-on-exception-tests.js:8]

-- Running test case: BreakOnAnyException.UncaughtExceptionInFinally
PASS: Should pause for exception.
PAUSE AT doThrow:8:22
Uncaught exception in test page: TestError [break-on-exception-tests.js:8]

-- Running test case: BreakOnAnyException.UncaughtExceptionThruNativeCode
PASS: Should pause for exception.
PAUSE AT doThrow:8:22
Uncaught exception in test page: TestError [break-on-exception-tests.js:8]

-- Running test case: BreakOnAnyException.CaughtException
PASS: Should pause for exception.
PAUSE AT doThrow:8:22
